Knappsack R2 Release

+++ knappsack-core/src/main/java/com/sparc/knappsack/components/events/UserInviteEvent.java
@@ -1,16 +1,13 @@
 package com.sparc.knappsack.components.events;
 
-import com.sparc.knappsack.components.entities.Domain;
-import com.sparc.knappsack.components.entities.Invitation;
-import com.sparc.knappsack.components.entities.User;
-import com.sparc.knappsack.components.services.DomainService;
+import com.sparc.knappsack.components.entities.*;
 import com.sparc.knappsack.components.services.EmailService;
+import com.sparc.knappsack.components.services.EmailServiceImpl;
 import com.sparc.knappsack.components.services.InvitationService;
 import com.sparc.knappsack.components.services.UserService;
 import com.sparc.knappsack.enums.DomainType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 
 @Component("userInviteEvent")
@@ -19,10 +16,6 @@ public class UserInviteEvent implements EventDelivery<Invitation> {
     @Qualifier("emailDeliveryService")
     @Autowired(required = true)
     private EmailService emailService;
-
-    @Qualifier("domainService")
-    @Autowired(required = true)
-    private DomainService domainService;
 
     @Qualifier("userService")
     @Autowired(required = true)
@@ -39,21 +32,25 @@ public class UserInviteEvent implements EventDelivery<Invitation> {
         if (invitation != null) {
 
             User fromUser = userService.getUserFromSecurityContext();
-            Domain domain = domainService.get(invitation.getDomainId(), invitation.getDomainType());
+            Domain domain = invitation.getDomain();
 
-            User invitee = userService.getByEmail(invitation.getEmail());
-            if(invitee != null) {
-                if(DomainType.GROUP.equals(invitation.getDomainType())) {
-                    userService.addUserToGroup(invitee, invitation.getDomainId(), invitation.getRole().getUserRole());
-                } else if(DomainType.ORGANIZATION.equals(invitation.getDomainType())) {
-                    userService.addUserToOrganization(invitee, invitation.getDomainId(), invitation.getRole().getUserRole());
+            if (domain != null) {
+
+                User invitee = userService.getByEmail(invitation.getEmail());
+                Long fromUserId = fromUser != null ? fromUser.getId() : null;
+                success = emailService.sendInvitationEmail(fromUserId, invitation.getId());
+                if(invitee != null) {
+                    if(DomainType.GROUP.equals(domain.getDomainType())) {
+                        userService.addUserToGroup(invitee, (Group) domain, invitation.getRole().getUserRole());
+                    } else if(DomainType.ORGANIZATION.equals(domain.getDomainType())) {
+                        userService.addUserToOrganization(invitee, (Organization) domain, invitation.getRole().getUserRole());
+                    }
+
+                    if (emailService instanceof EmailServiceImpl) {
+                        invitationService.delete(invitation.getId());
+                    }
                 }
-
-                invitationService.delete(invitation.getId());
-
             }
-
-            success = emailService.sendInvitationEmail(fromUser, invitation.getEmail(), (domain == null ? "" : domain.getName()), invitation.getDomainType());
         }
 
         return success;
